+++ workout_diary/src/database_connection.py
@@ -8,5 +8,3 @@ connection.row_factory = sqlite3.Row
 
 def get_database_connection():
     return connection
-
-