+++ workout_diary/src/main.py
@@ -1,11 +1,12 @@
-from ui.ui import Ui
+from ui.ui import UI
 from entities.user import User
 from entities.workout import Workout
 from services.workout_service import WorkoutService
 
 
 def main():
-    Ui.main_window()
+    UI.start()
+    
 
 
 if __name__ == "__main__":