+++ workout_diary/src/ui/ui.py
@@ -1,38 +1,67 @@
-import tkinter as tk
+from tkinter import Tk, ttk, constants
 from repositories.user_repository import UserRepository
+from ui.new_user_window import NewUserWindow
 
-class Ui:
-    def __init__(self):
-        pass
+class UI:
+    def __init__(self, root):
+        self._root = root
+        self._login_username = None
+        self._login_password = None
+        self._current_view = None
+    
+    def start(self):
+        heading_label = ttk.Label(master=self._root, text="Welcome to Workout Diary!")
+        
+        login_label = ttk.Label(master=self._root, text="Login:")
 
-    def main_window():
-        window = tk.Tk()
-        window.title("Workout diary")
-        window.geometry("1000x300")
+        username_label = ttk.Label(master=self._root, text="Username")
+        self._username_entry = ttk.Entry(master=self._root)
 
-        T = tk.Text(window, height=5, width=62)
-        l = tk.Label(
-            window, text="You have just started Ville Tonteri's first program - Workout Diary!")
-        l.config(font=("Courier", 14))
+        password_label = ttk.Label(master=self._root, text="Password")
+        self._login_password = ttk.Entry(master=self._root)
 
-        Fact = """Program doesn't yet have anything to show, except this window. 
-        More to follow on coming weeks - stay tuned!"""
+        button_stop_application = ttk.Button(master=self._root, text='Stop application',
+                            width=15, command=self._root.destroy)
 
-        button_stop_application = tk.Button(window, text='Stop application',
-                           width=15, command=window.destroy)
+        button_create_user = ttk.Button(master=self._root, text="Create New User",
+                            width=15, command=UI._create_user_window(self))
 
-        button_create_user = tk.Button(window, text="Create New User",
-                            width=15, command=Ui._create_user_window)
+        button_login_user = ttk.Button(master=self._root, text="Login",
+                            width=15, command=UI._main_window)
+        
+        heading_label.grid(row=0, column=0, columnspan=2, sticky=(constants.E, constants.W), padx=5, pady=5)
 
-        button_stop_application.pack()
-        l.pack()
-        T.pack()
-        T.place(x=280, y=150)
-        l.place(x=120, y=80)
-        button_stop_application.place(x=450, y=250)
-        button_create_user.place(x=450, y=200)
-        T.insert(tk.END, Fact)
-        window.mainloop()
+        login_label.grid(row=1, column=0, sticky=constants.W, padx=5, pady=5)
+
+        username_label.grid(row=2, column=0, padx=5, pady=5)
+        self._username_entry.grid(row=2, column=1, sticky=(constants.E, constants.W), padx=5, pady=5)
+
+        password_label.grid(row=3, column=0, padx=5, pady=5)
+        self._login_password.grid(row=3, column=1, sticky=(constants.E, constants.W), padx=5, pady=5)
+
+        button_login_user.grid(row=4, column=0, columnspan=2, padx=5, pady=5)
+
+        button_create_user.grid(row=5, column=0, columnspan=2, padx=5, pady=5)
+        
+        button_stop_application.grid(row=6, column=0, columnspan=2, padx=5, pady=5)
+
+        self._root.grid_columnconfigure(1, weight = 1)
+
 
     def _create_user_window(self):
+        self._current_view = NewUserWindow(self._root)
+        self._current_view.pack()
         pass
+
+    def _main_window(self):
+
+        pass
+
+window = Tk()
+window.title("WORKOUT DIARY APPLICATION")
+window.geometry("350x230")
+
+
+ui = UI(window)
+ui.start()
+window.mainloop()